Pull charges summarizer into separate class file.
